// code/src/pages/licensePlan/components/PaymentOptions/paymentOptions.enums.ts
export const enum PAYMENT_OPTIONS {
  NEW_CARD = 'newCard',
  SAVED_CARD = 'savedCard',
  OTHER = 'other'
}

export const enum PAYMENT_OPTIONS_VALUES {
  PAYMENT_METHOD = 'paymentMethod',
  CARD_NUMBER = 'cardNumber',
  CARD_EXPIRY_MONTH = 'cardExpiryMonth',
  CARD_EXPIRY_YEAR = 'cardExpiryYear',
  CARD_CVC = 'cardCvc',
  CARD_NAME = 'cardName',
  CARD_POSTAL_CODE = 'cardPostalCode',
  CARD_ADDRESS_LINE1 = 'cardAddressLine1',
  CARD_CITY = 'cardCity',
  CARD_COUNTRY = 'cardCountry'
}